# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'The Financial Express API'
  contact:
    name: 'Eliyas Hossain'
    url: 'https://batterylowinteractive.com'
    email: eliyas.batterylowinteractive@gmail.com
  version: 1.0.0
servers:
  -
    url: /admin
    description: 'Admin API base url'
paths:
...
  '/api/{locale}/category/world/home':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world category'
      description: 'Get posts by world category'
      operationId: worldCategoryApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
  '/api/{locale}/category/world/asia':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world asia category'
      description: 'Get posts by world asia category'
      operationId: worldAsiaCategoryApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
  '/api/{locale}/category/world/america':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world america category'
      description: 'Get posts by world america category'
      operationId: worldAmericaCategoryApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
  '/api/{locale}/category/world/europe':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world europe category'
      description: 'Get posts by world europe category'
      operationId: worldEuropeCategoryApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
  '/api/{locale}/category/world/africa':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world africa category'
      description: 'Get posts by world africa category'
      operationId: worldAfricaCategoryApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
  '/api/{locale}/category/world/more':
    get:
      tags:
        - WorldCategoryApi
      summary: 'Get posts by world more'
      description: 'Get posts by world more by paginate'
      operationId: worldCategoryMoreApiIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostApi'
                type: object
        400:
          $ref: '#/components/responses/400'
        5XX:
          $ref: '#/components/responses/500'
      servers:
        -
          url: /
          description: 'API base url'
...
  '/{locale}/category':
    get:
      tags:
        - Category
      summary: 'Get category'
      description: 'Get category list'
      operationId: categoryIndex
      parameters:
        -
          $ref: '#/components/parameters/locale'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryResponse'
                type: object
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        5XX:
          $ref: '#/components/responses/500'
      security:
        -
          sanctum: []
    post:
      tags:
        - Category
      summary: 'Store category'
      description: 'Create category'
      operationId: categoryStore
      parameters:
        -
          $ref: '#/components/parameters/locale'
      requestBody:
        description: 'Please enter valid information'
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Success
                  categoryId:
                    type: integer
                    example: '1'
                type: object
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        5XX:
          $ref: '#/components/responses/500'
      security:
        -
          sanctum: []
...
  '/{locale}/category/{id}':
    get:
      tags:
        - Category
      summary: 'Get single category'
      description: 'Get category'
      operationId: categoryShow
      parameters:
        -
          $ref: '#/components/parameters/locale'
        -
          $ref: '#/components/parameters/id'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                properties:
                  data:
                    $ref: '#/components/schemas/Category'
                type: object
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
        5XX:
          $ref: '#/components/responses/500'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Category
      summary: 'Delete category'
      description: 'Delete category'
      operationId: categoryDelete
      parameters:
        -
          $ref: '#/components/parameters/locale'
        -
          $ref: '#/components/parameters/id'
      responses:
        200:
          $ref: '#/components/responses/200'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        5XX:
          $ref: '#/components/responses/500'
      security:
        -
          sanctum: []
    patch:
      tags:
        - Category
      summary: 'Update category'
      description: 'Update category'
      operationId: categoryUpdate
      parameters:
        -
          $ref: '#/components/parameters/locale'
        -
          $ref: '#/components/parameters/id'
      requestBody:
        description: 'Please enter valid information'
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        200:
          $ref: '#/components/responses/200'
        400:
          $ref: '#/components/responses/400'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        5XX:
          $ref: '#/components/responses/500'
      security:
        -
          sanctum: []
...
components:
  schemas:
...
    Category:
      title: Category
      description: 'Category model'
      required:
        - name
        - slug
        - status
        - ordering
      properties:
        id:
          title: id
          description: 'Category ID'
          type: integer
          format: int64
          readOnly: true
        status:
          title: status
          description: 'Check, if resource active or inactive'
          type: boolean
          format: int64
        name:
          title: name
          description: 'Category name'
          type: string
          format: string
        slug:
          title: slug
          description: 'Category slug'
          type: string
          format: string
        color:
          title: color
          description: 'Category color'
          type: string
          format: string
        ordering:
          title: ordering
          description: 'Order number'
          type: integer
          format: int64
        parent_id:
          title: parent_id
          description: 'Parent ID'
          type: integer
          format: int64
        meta_title:
          title: meta_title
          description: 'Category meta title'
          type: string
          format: string
        meta_description:
          title: meta_description
          description: 'Category meta description'
          type: string
          format: string
        meta_image:
          title: meta_image
          description: 'Category meta image'
          type: string
          format: binary
      type: object
    CategoryChildResponse:
      title: CategoryChildResponse
      description: 'Category children response'
      properties:
        id:
          title: id
          description: 'Category ID'
          type: integer
          format: int64
        name:
          title: name
          description: 'Category name'
          type: string
          format: string
        slug:
          title: slug
          description: 'Category slug'
          type: string
          format: string
        color:
          title: color
          description: 'Category color'
          type: string
          format: string
      type: object
    CategoryCommonApi:
      title: CategoryCommonApi
      description: 'Category common response for API'
      properties:
        id:
          title: id
          description: 'Category ID'
          type: integer
          format: int64
        name:
          title: name
          description: 'Category name'
          type: string
          format: string
        slug:
          title: slug
          description: 'Category slug'
          type: string
          format: string
        color:
          title: color
          description: 'Category color'
          type: string
          format: string
      type: object
    CategoryResponse:
      title: CategoryResponse
      description: 'Category list response'
      properties:
        id:
          title: id
          description: 'Category ID'
          type: integer
          format: int64
        name:
          title: name
          description: 'Category name'
          type: string
          format: string
        slug:
          title: slug
          description: 'Category slug'
          type: string
          format: string
        color:
          title: color
          description: 'Category color'
          type: string
          format: string
        posts_count:
          title: posts_count
          description: 'Category posts count'
          type: integer
          format: int64
        children:
          title: Children
          description: children
          type: array
          items:
            $ref: '#/components/schemas/CategoryChildResponse'
      type: object
...
    PostApi:
      title: PostApi
      description: 'PostResponse model'
      properties:
        id:
          title: id
          description: 'Post ID'
          type: integer
          format: int64
          readOnly: true
        title:
          title: title
          description: 'Post title'
          type: string
          format: string
        slug:
          title: slug
          description: 'Post slug'
          type: string
          format: string
        excerpt:
          title: excerpt
          description: 'Post excerpt'
          type: string
          format: string
        image:
          title: image
          description: 'Post image'
          type: string
          format: binary
        caption:
          title: caption
          description: 'Post image caption'
          type: string
          format: string
        datetime:
          title: datetime
          description: 'Post datetime'
          type: string
          format: date-time
        category:
          $ref: '#/components/schemas/CategoryCommonApi'
      type: object
...
  responses:
    200:
      description: Success
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: Success
            type: object
    400:
      description: 'Bad Request'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Something weng wrong, please try again later.'
            type: object
    401:
      description: Unauthenticated
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Please sign in to use this service.'
            type: object
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'This action is unauthorized.'
            type: object
    404:
      description: 'Not Found'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Resource Not Found.'
            type: object
    500:
      description: 'Internal Server Error'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Internal Server Error.'
            type: object
  parameters:
    locale:
      name: locale
      in: path
      description: 'App locale'
      required: true
      schema:
        type: string
      example: en
    id:
      name: id
      in: path
      description: 'Model ID'
      required: true
      schema:
        type: integer
        format: int64
      example: '1'
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Laravel Sanctum token authentication'
      scheme: bearer
      bearerFormat: JWT
security:
  - []